Collapse base and derived class
